FragmentContentExtractor unit tests
