+++ knotx-junit/pom.xml
@@ -54,6 +54,16 @@
       <artifactId>system-rules</artifactId>
       <scope>provided</scope>
     </dependency>
+    <dependency>
+      <groupId>org.mockito</groupId>
+      <artifactId>mockito-all</artifactId>
+      <scope>provided</scope>
+    </dependency>
+    <dependency>
+      <groupId>com.googlecode.zohhak</groupId>
+      <artifactId>zohhak</artifactId>
+      <scope>provided</scope>
+    </dependency>
   </dependencies>
 
   <build>
