+++ knotx-knot/knotx-knot-handlebars/pom.xml
@@ -79,6 +79,10 @@
       <artifactId>vertx-unit</artifactId>
     </dependency>
     <dependency>
+      <groupId>com.googlecode.zohhak</groupId>
+      <artifactId>zohhak</artifactId>
+    </dependency>
+    <dependency>
       <groupId>io.knotx</groupId>
       <artifactId>knotx-junit</artifactId>
       <version>${project.version}</version>
