+++ knotx-knot/knotx-knot-handlebars/src/main/java/io/knotx/knot/templating/impl/FragmentContentExtractor.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright (C) 2016 Cognifide Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.knotx.knot.templating.impl;
+
+import io.knotx.dataobjects.Fragment;
+import org.jsoup.Jsoup;
+import org.jsoup.nodes.Document;
+import org.jsoup.nodes.Element;
+
+final class FragmentContentExtractor {
+
+  private FragmentContentExtractor() {
+    //util class
+  }
+
+  static String getUnwrappedContent(Fragment fragment) {
+    Document document = Jsoup.parseBodyFragment(fragment.content());
+    Element scriptTag = document.body().child(0);
+    return scriptTag.unwrap().toString();
+  }
+}
