+++ knotx-knot/knotx-knot-handlebars/src/test/resources/big_snippet-expected_content.txt
@@ -1,0 +1,16 @@
+<div class="col-md-4">
+  <h2>Snippet1 - {{_result.message}}</h2>
+  <div>Snippet1 - {{_result.body.a}}</div>
+  <div>{{#bold}}Custom Handlebars helper{{/bold}}</div>
+</div>
+<div class="col-md-4">
+  <h2>Snippet1 - {{second._result.message}}</h2>
+  <div>Snippet1 - {{second._result.body.a}}</div>
+  {{#string_equals second._response.statusCode "200"}}
+  <div>Success! Status code : {{second._response.statusCode}}</div>
+  {{/string_equals}}
+</div>
+<div class="col-md-4">
+  <h2>Snippet1 - {{third._result.message}}</h2>
+  <div>Snippet1 - {{third._result.body.a}}</div>
+</div>
