+++ knotx-knot/knotx-knot-handlebars/src/test/resources/simple_snippet-expected_content.txt
@@ -1,0 +1,1 @@
+<h2>{{_result.message}}</h2>
