+++ knotx-knot/knotx-knot-handlebars/src/test/resources/simple_snippet.txt
@@ -1,0 +1,7 @@
+<script
+  data-knotx-knots="services"
+  data-knotx-service="first-service"
+  data-knotx-params='{"path":"/overridden/path"}'
+  type="text/knotx-snippet">
+  <h2>{{_result.message}}</h2>
+</script>
